- Added ScraperAlbumInformation to project file - Added CoreConfigProperties - MetaFileWriter now uses CoreConfigProperties to determine metaFilePath

+++ src/NoisedCore/Media/MetaFileWriter.cs
@@ -1,8 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using Noised.Core.Config;
+using Noised.Core.etc;
 using Noised.Core.IOC;
 
 namespace Noised.Core.Media
@@ -19,7 +18,9 @@ namespace Noised.Core.Media
         /// </summary>
         public MetaFileWriter()
         {
-            metaFilePath = IocContainer.Get<IConfig>().GetProperty("noised.core.metafiles");
+            metaFilePath = IocContainer.Get<IConfig>().GetProperty(CoreConfigProperties.MetaFileDirectory);
+
+            CreateDirectoryIfNotExists(metaFilePath);
         }
 
         #region Implementation of IMetaFileWriter
@@ -30,38 +31,36 @@ namespace Noised.Core.Media
         /// <param name="metaFile">The MetaFile which should be written to the disk</param>
         public void WriteMetaFileToDisk(MetaFile metaFile)
         {
-            string filename = GetFileName(metaFile);
+            // always append the artist
+            string fullPath = Path.Combine(metaFilePath, metaFile.Artist);
 
-            string fullPath = Path.Combine(metaFilePath, filename);
+            // if the metafile has an album attach the album
+            if (!String.IsNullOrWhiteSpace(metaFile.Album))
+                fullPath = Path.Combine(fullPath, metaFile.Album);
 
+            // create the directory if it does not already exist
+            CreateDirectoryIfNotExists(fullPath);
+
+            // attach the filename
+            fullPath = Path.Combine(fullPath, metaFile.OriginalFilename);
+
+            // write only if file not already exists
+            if (File.Exists(fullPath))
+                return;
+
+            // change URI to local URI
             metaFile.Uri = new Uri(fullPath);
+
+            // write file to disk
             File.WriteAllBytes(fullPath, metaFile.Data);
         }
 
         #endregion
 
-        /// <summary>
-        /// Gets a filename and path for a MetaFile
-        /// </summary>
-        /// <param name="metaFile">The MetaFile for which a path and filename should be created</param>
-        /// <returns>A valid pathh and filename</returns>
-        private string GetFileName(MetaFile metaFile)
+        private static void CreateDirectoryIfNotExists(string dir)
         {
-            string path = Path.Combine(metaFilePath, metaFile.Artist, metaFile.Album);
-
-            List<string> files = Directory.EnumerateFileSystemEntries(path, metaFile.Type + "_*").ToList();
-
-            int i = 0;
-            string fileName;
-
-            do
-            {
-                fileName = metaFile.Type + i + metaFile.Extension;
-                i++;
-            }
-            while (files.Contains(fileName));
-
-            return fileName;
+            if (!Directory.Exists(dir))
+                Directory.CreateDirectory(dir);
         }
     }
 }
